terraform_code/backend_files/requirements.txt
Flask==2.2.5
mysql-connector-python==8.3.0
boto3==1.34.79
